Refactor Form
