+++ src/components/App/App.jsx
@@ -4,10 +4,10 @@ import "./App.css";
 import axios from "axios";
 import { useEffect, useState } from "react";
 import ShoppingList from "../ShoppingList/ShoppingList.jsx";
+import ShoppingListForm from "../ShoppingItemForm/ShoppingItemForm.jsx";
 
 function App() {
   const [shoppingList, setShoppingList] = useState([]);
-  let [shoppingFormItem, setShoppingFormItem] = useState("Mahad");
   // let[shoppingFormQuantity, setShoppingFormQuantity] = useState("");
   // let[shoppingFormUnit, setShoppingFormUnit] = useState("");
   // let[shoppingFormBuy, setShoppingFormBuy] = useState("");
@@ -33,42 +33,12 @@ function App() {
 
   //! POST FUNCTION
 
-  const addItem = () => {
-    axios
-      .post("/api/shoppingList", { item: shoppingFormItem })
-      .then((response) => {
-        setShoppingFormItem("");
-        fetchShoppingList();
-      })
-      .catch((err) => {
-        alert("Error Adding Item");
-        console.log(err);
-      });
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    //! We need to add more input variables below
-    if (shoppingFormItem) {
-      addItem();
-    } else {
-      alert("We Need an Item");
-    }
-  };
 
   return (
     <div className="App">
       <Header />
       <main>
-        <form onSubmit={handleSubmit}>
-          <input
-            type="text"
-            placeholder="Item"
-            value={shoppingFormItem}
-            onChange={(evt) => setShoppingFormItem(evt.target.value)}
-          ></input>
-          <button type="submit">Add Item</button>
-        </form>
+        <ShoppingListForm />
         <ShoppingList shoppingList={shoppingList} />
       </main>
     </div>
